[REFACTOR] Move small functions to helper.hpp
